observability: fix AV.4 boundary findings
