+++ crates/sc-observability/src/lib.rs
@@ -7,7 +7,6 @@
 //! - socket error-code constants for the `log-event` contract
 
 use agent_team_mail_core::logging_event::{LogEventV1, ValidationError};
-use chrono::{SecondsFormat, Utc};
 use std::fs::{self, OpenOptions};
 use std::io::Write;
 use std::path::{Path, PathBuf};
@@ -16,7 +15,10 @@ use std::sync::{Arc, Mutex, OnceLock};
 use std::time::Duration;
 use thiserror::Error;
 
+mod health;
 mod otlp_adapter;
+
+pub use health::{OtelHealthSnapshot, OtelLastError, current_otel_health};
 
 pub const DEFAULT_QUEUE_CAPACITY: usize = 4096;
 pub const DEFAULT_MAX_EVENT_BYTES: usize = 64 * 1024;
@@ -239,204 +241,6 @@ impl OtelPipeline {
     }
 }
 
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, PartialEq, Eq, Default)]
-pub struct OtelLastError {
-    pub code: Option<String>,
-    pub message: Option<String>,
-    pub at: Option<String>,
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, PartialEq, Eq)]
-pub struct OtelHealthSnapshot {
-    pub schema_version: String,
-    pub enabled: bool,
-    pub collector_endpoint: Option<String>,
-    pub protocol: String,
-    pub collector_state: String,
-    pub local_mirror_state: String,
-    pub local_mirror_path: String,
-    pub debug_local_export: bool,
-    pub debug_local_state: String,
-    pub last_error: OtelLastError,
-}
-
-impl Default for OtelHealthSnapshot {
-    fn default() -> Self {
-        Self {
-            schema_version: "v1".to_string(),
-            enabled: true,
-            collector_endpoint: None,
-            protocol: OTEL_PROTOCOL_HTTP.to_string(),
-            collector_state: "not_configured".to_string(),
-            local_mirror_state: "healthy".to_string(),
-            local_mirror_path: String::new(),
-            debug_local_export: false,
-            debug_local_state: "disabled".to_string(),
-            last_error: OtelLastError::default(),
-        }
-    }
-}
-
-#[derive(Debug, Default)]
-struct OtelRuntimeHealth {
-    collector_error: Option<String>,
-    collector_error_at: Option<String>,
-    collector_success_at: Option<String>,
-    local_mirror_error: Option<String>,
-    local_mirror_error_at: Option<String>,
-    local_mirror_success_at: Option<String>,
-    debug_local_error: Option<String>,
-    debug_local_error_at: Option<String>,
-    debug_local_success_at: Option<String>,
-}
-
-fn otel_runtime_health_slot() -> &'static Mutex<OtelRuntimeHealth> {
-    static OTEL_RUNTIME_HEALTH: OnceLock<Mutex<OtelRuntimeHealth>> = OnceLock::new();
-    OTEL_RUNTIME_HEALTH.get_or_init(|| Mutex::new(OtelRuntimeHealth::default()))
-}
-
-fn now_rfc3339() -> String {
-    Utc::now().to_rfc3339_opts(SecondsFormat::Secs, true)
-}
-
-fn note_export_success(kind: OtelExporterKind) {
-    let mut health = otel_runtime_health_slot()
-        .lock()
-        .expect("sc-observability otel health lock poisoned");
-    let now = now_rfc3339();
-    match kind {
-        OtelExporterKind::Collector => {
-            health.collector_error = None;
-            health.collector_error_at = None;
-            health.collector_success_at = Some(now);
-        }
-        OtelExporterKind::LocalMirror => {
-            health.local_mirror_error = None;
-            health.local_mirror_error_at = None;
-            health.local_mirror_success_at = Some(now);
-        }
-        OtelExporterKind::DebugLocal => {
-            health.debug_local_error = None;
-            health.debug_local_error_at = None;
-            health.debug_local_success_at = Some(now);
-        }
-    }
-}
-
-fn note_export_failure(kind: OtelExporterKind, err: &OtelError) {
-    let mut health = otel_runtime_health_slot()
-        .lock()
-        .expect("sc-observability otel health lock poisoned");
-    let now = now_rfc3339();
-    match kind {
-        OtelExporterKind::Collector => {
-            health.collector_error = Some(err.to_string());
-            health.collector_error_at = Some(now);
-        }
-        OtelExporterKind::LocalMirror => {
-            health.local_mirror_error = Some(err.to_string());
-            health.local_mirror_error_at = Some(now);
-        }
-        OtelExporterKind::DebugLocal => {
-            health.debug_local_error = Some(err.to_string());
-            health.debug_local_error_at = Some(now);
-        }
-    }
-}
-
-fn health_state(
-    configured: bool,
-    success_at: &Option<String>,
-    error_at: &Option<String>,
-) -> String {
-    if !configured {
-        return "disabled".to_string();
-    }
-    match (success_at, error_at) {
-        (Some(success), Some(error)) => {
-            if error > success {
-                "degraded".to_string()
-            } else {
-                "healthy".to_string()
-            }
-        }
-        (Some(_), None) => "healthy".to_string(),
-        (None, Some(_)) => "degraded".to_string(),
-        (None, None) => "configured".to_string(),
-    }
-}
-
-pub fn current_otel_health(log_path: &Path) -> OtelHealthSnapshot {
-    let config = OtelConfig::from_env();
-    let local_mirror_path = default_otel_path(log_path);
-    let health = otel_runtime_health_slot()
-        .lock()
-        .expect("sc-observability otel health lock poisoned");
-
-    let collector_configured = config.enabled
-        && config
-            .endpoint
-            .as_deref()
-            .is_some_and(|value| !value.trim().is_empty());
-    let debug_local_configured = config.enabled && config.debug_local_export;
-    let local_mirror_configured = config.enabled;
-
-    let last_error = if health.collector_error.is_some() {
-        OtelLastError {
-            code: Some("COLLECTOR_EXPORT_FAILED".to_string()),
-            message: health.collector_error.clone(),
-            at: health.collector_error_at.clone(),
-        }
-    } else if health.local_mirror_error.is_some() {
-        OtelLastError {
-            code: Some("LOCAL_MIRROR_EXPORT_FAILED".to_string()),
-            message: health.local_mirror_error.clone(),
-            at: health.local_mirror_error_at.clone(),
-        }
-    } else if health.debug_local_error.is_some() {
-        OtelLastError {
-            code: Some("DEBUG_LOCAL_EXPORT_FAILED".to_string()),
-            message: health.debug_local_error.clone(),
-            at: health.debug_local_error_at.clone(),
-        }
-    } else {
-        OtelLastError::default()
-    };
-
-    let collector_state = if !config.enabled {
-        "disabled".to_string()
-    } else if !collector_configured {
-        "not_configured".to_string()
-    } else {
-        health_state(
-            true,
-            &health.collector_success_at,
-            &health.collector_error_at,
-        )
-    };
-
-    OtelHealthSnapshot {
-        schema_version: "v1".to_string(),
-        enabled: config.enabled,
-        collector_endpoint: config.endpoint.clone(),
-        protocol: config.protocol.clone(),
-        collector_state,
-        local_mirror_state: health_state(
-            local_mirror_configured,
-            &health.local_mirror_success_at,
-            &health.local_mirror_error_at,
-        ),
-        local_mirror_path: local_mirror_path.to_string_lossy().into_owned(),
-        debug_local_export: config.debug_local_export,
-        debug_local_state: health_state(
-            debug_local_configured,
-            &health.debug_local_success_at,
-            &health.debug_local_error_at,
-        ),
-        last_error,
-    }
-}
-
 fn export_otel_with_retry(
     event: &LogEventV1,
     config: &OtelConfig,
@@ -459,11 +263,11 @@ fn export_otel_with_retry(
         let mut any_failed = false;
         for exporter in exporters {
             if let Err(err) = exporter.export(&record) {
-                note_export_failure(exporter.kind(), &err);
+                health::note_export_failure(exporter.kind(), &err);
                 any_failed = true;
                 last_err = Some(err);
             } else {
-                note_export_success(exporter.kind());
+                health::note_export_success(exporter.kind());
             }
         }
         if !any_failed {
@@ -499,10 +303,10 @@ pub fn export_otel_best_effort(
     let mut backoff = config.initial_backoff_ms;
     loop {
         if exporter.export(&record).is_ok() {
-            note_export_success(exporter.kind());
+            health::note_export_success(exporter.kind());
             return;
         }
-        note_export_failure(
+        health::note_export_failure(
             exporter.kind(),
             &OtelError::ExportFailed("best-effort exporter failed".to_string()),
         );
@@ -1096,7 +900,7 @@ fn rotate_log_files(base: &Path, max_files: u32) -> Result<(), LoggerError> {
     Ok(())
 }
 
-fn default_otel_path(log_path: &Path) -> PathBuf {
+pub(crate) fn default_otel_path(log_path: &Path) -> PathBuf {
     let mut otel_path = log_path.to_path_buf();
     let stem = log_path
         .file_stem()
